+++ src/comm/ingest.rs
@@ -55,7 +55,7 @@ const CHANNEL_CLOSE_MESSAGE: &[u8; 12] = b"channel done";
 const CHANNEL_CLOSE_TIMESTAMP: i64 = -1;
 const NO_TIMESTAMP: i64 = 0;
 const SENSOR_INTERVAL: u64 = 60 * 60 * 24;
-const INGEST_VERSION_REQ: &str = ">=0.26.0-alpha.1,<0.26.0-alpha.2";
+const INGEST_VERSION_REQ: &str = ">=0.26.0-alpha.2,<0.26.0-alpha.3";
 
 type SensorInfo = (String, DateTime<Utc>, ConnState, bool);