+++ src/comm/publish/tests.rs
@@ -63,7 +63,7 @@ fn get_token() -> &'static Mutex<u32> {
 }
 
 const CA_CERT_PATH: &str = "tests/certs/ca_cert.pem";
-const PROTOCOL_VERSION: &str = "0.26.0-alpha.1";
+const PROTOCOL_VERSION: &str = "0.26.0-alpha.2";
 
 const NODE1_CERT_PATH: &str = "tests/certs/node1/cert.pem";
 const NODE1_KEY_PATH: &str = "tests/certs/node1/key.pem";
@@ -276,11 +276,9 @@ fn gen_http_raw_event() -> Vec<u8> {
         content_encoding: String::new(),
         content_type: String::new(),
         cache_control: String::new(),
-        orig_filenames: Vec::new(),
-        orig_mime_types: Vec::new(),
-        resp_filenames: Vec::new(),
-        resp_mime_types: Vec::new(),
-        post_body: Vec::new(),
+        filenames: Vec::new(),
+        mime_types: Vec::new(),
+        body: Vec::new(),
         state: String::new(),
     };